Address review: lossless encoding schema, unsigned NTP64, delegate
- Encoding schema is raw bytes on the wire (Option<ZSlice>, deserialized
  with no UTF-8 validation), so a non-UTF-8 schema can arrive from any
  peer. Return it losslessly: encoding_get_schema and EncodingStruct.schema
  are now Option<Vec<u8>> instead of from_utf8(..).ok() (which mapped a
  present non-UTF-8 schema to None, indistinguishable from absent).
- Timestamp.ntp64 is now u64, not i64: NTP64 is unsigned and current-era
  values set the high bit, which the i64 cast rendered negative. The
  sample constructors' timestamp_ntp64 input is u64 to match.
- EncodingStruct::from delegates to encoding_get_id / encoding_get_schema
  so each field has a single definition (one source of truth).
- Regression tests: a non-UTF-8 schema round-trips; an NTP64 value above
  i64::MAX is preserved unsigned.

+++ src/base/sample/mod.rs
@@ -24,7 +24,7 @@ pub fn sample_new_put(
     key_expr: KeyExpr,
     payload: ZBytes,
     encoding: Option<&Encoding>,
-    timestamp_ntp64: Option<i64>,
+    timestamp_ntp64: Option<u64>,
     attachment: Option<ZBytes>,
     congestion_control: Option<CongestionControl>,
     priority: Option<Priority>,
@@ -36,10 +36,7 @@ pub fn sample_new_put(
         builder = builder.encoding(enc.clone());
     }
     if let Some(ntp) = timestamp_ntp64 {
-        builder = builder.timestamp(zenoh::time::Timestamp::new(
-            NTP64(ntp as u64),
-            TimestampId::rand(),
-        ));
+        builder = builder.timestamp(zenoh::time::Timestamp::new(NTP64(ntp), TimestampId::rand()));
     }
     if let Some(att) = attachment {
         builder = builder.attachment(att);
@@ -69,7 +66,7 @@ pub fn sample_new_put(
 #[prebindgen]
 pub fn sample_new_delete(
     key_expr: KeyExpr,
-    timestamp_ntp64: Option<i64>,
+    timestamp_ntp64: Option<u64>,
     attachment: Option<ZBytes>,
     congestion_control: Option<CongestionControl>,
     priority: Option<Priority>,
@@ -78,10 +75,7 @@ pub fn sample_new_delete(
 ) -> Sample {
     let mut builder = SampleBuilder::delete(key_expr);
     if let Some(ntp) = timestamp_ntp64 {
-        builder = builder.timestamp(zenoh::time::Timestamp::new(
-            NTP64(ntp as u64),
-            TimestampId::rand(),
-        ));
+        builder = builder.timestamp(zenoh::time::Timestamp::new(NTP64(ntp), TimestampId::rand()));
     }
     if let Some(att) = attachment {
         builder = builder.attachment(att);
